% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_dist.R
\name{boot_dist}
\alias{boot_dist}
\title{boot_dist}
\usage{
boot_dist(df, n_boot = 250)
}
\arguments{
\item{df}{input data frame. should have columns \code{x}, \code{y}, and \code{delta}}

\item{n_boot}{number of bootstrap samples}
}
\value{
a dataframe containing correlations and the variance ratio for each
bootstrap sample
}
\description{
For a given dataframe \code{df}, generate \code{n_boot} bootstrap samples. For each
bootstrap sample, compute correlations between x and delta, and between x
and y. Slso compute the variance ratio.
}
\examples{

boot_dist(real_dfs,n_boot = 1)
}
