% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_from_resamp.R
\name{ci_from_resamp}
\alias{ci_from_resamp}
\title{ci_from_resamp}
\usage{
ci_from_resamp(df, method = "quant", n)
}
\arguments{
\item{df}{input dataframe. Contains estimates of correlations and k based
on repeated sampling.}

\item{method}{method to derive CI. For \code{quant}, uses quantiles. For
\code{t_quant}, uses a t-distribution adjusted quantile.}

\item{n}{See previous.}
}
\value{

}
\description{
Input is a dataframe containing results of a resampling analysis to look
at correlations between x and y, and between x and delta, as well as the
variance ratio. This returns confidence intervals for those values based on
the resampling results.
}
\examples{

ci_from_resamp(boot_analysis[[2]][[2]], method = "quant", 1)
}
