% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const_tobit_ll.R
\name{const_tobit_ll}
\alias{const_tobit_ll}
\title{const_tobit_ll
This function only takes a lower OR upper limit}
\usage{
const_tobit_ll(par, X, y, ul = -Inf, ll = Inf)
}
\arguments{
\item{ll}{}
}
\value{

}
\description{
const_tobit_ll
This function only takes a lower OR upper limit
}
\examples{

}
