% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_xd.R
\name{cor_xd}
\alias{cor_xd}
\title{cor_xd}
\usage{
cor_xd(k, cor_xy)
}
\value{
plots
}
\description{
function computes cor_xd from k and cor_xy. used especially for creating plots.
}
\examples{

cor_xd(0.1, -1)
}
