% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_const.R
\name{fit_const}
\alias{fit_const}
\title{fit_const
This function fit the constant model}
\usage{
fit_const(formula, data, thresh = 66)
}
\arguments{
\item{thresh}{}
}
\value{

}
\description{
fit_const
This function fit the constant model
}
\examples{
fit_const(y_obs ~ x, data = df)
}
