% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_smooth.R
\name{fit_smooth}
\alias{fit_smooth}
\title{fit_smooth}
\usage{
fit_smooth(formula, data, thresh = 66)
}
\arguments{
\item{thresh}{}
}
\value{

}
\description{
fit_smooth
}
\examples{
fit_smooth(y_obs ~ x, data = df)
}
