% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_fun.R
\name{hclust_fun}
\alias{hclust_fun}
\title{hclust_fun}
\usage{
hclust_fun(df, k)
}
\arguments{
\item{df}{dataframe to cluster; should contain only variables that are used
in the clustering process}

\item{k}{number of clusters to identify}
}
\value{
vector of cluster assignments
}
\description{
This function is a wrapper for a few lines that implement a clustering
analysis. Given an input dataframe and number of clusters, this will conduct
hierarchical cluster using choices that mimic those in Winters et al and
others. Clustering is performed on all variables in the input dataframe;
often specific variables are chosen at the outset (e.g. x and y).
}
\examples{

hclust_fun(winters_df \%>\% select(x, y), 2)
}
