% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fit_const.R
\name{predict.fit_const}
\alias{predict.fit_const}
\title{predict.fit_const
This function predict the constant model}
\usage{
\method{predict}{fit_const}(object, data.new = NULL)
}
\arguments{
\item{data.new}{}
}
\value{

}
\description{
predict.fit_const
This function predict the constant model
}
