% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fit_prr.R
\name{predict.fit_prr}
\alias{predict.fit_prr}
\title{predict.fit_prr
This function predict the proportion recovery model}
\usage{
\method{predict}{fit_prr}(object, data.new = NULL)
}
\arguments{
\item{data.new}{}
}
\value{

}
\description{
predict.fit_prr
This function predict the proportion recovery model
}
