% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_recov_data.R
\name{random_recov_data}
\alias{random_recov_data}
\title{random_recov_data}
\usage{
random_recov_data(seed, start_df)
}
\arguments{
\item{seed}{seed for reproducibility}

\item{start_df}{data set that is used as the basis for data simulated under
random recovery.}
}
\value{
dataframe containing randomly generated data
}
\description{
Given a starting dataset, this function extracts observed baseline values
and, given these, simulates follow-up values between the baseline and 66
(the max allowed value) using a uniform distribution. The mechanism is
intended to mimic the random recovery process implemented in various papers.
This function is used in the context of clustering analysis and simulations.
}
\examples{

random_recov_data(1, winters_df)
}
