% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_dfs.R
\docType{data}
\name{real_dfs}
\alias{real_dfs}
\title{Combined public dataframe

A dataset containing the information to do analysis on
correlations between x and y, and between x and delta}
\format{
A data frame with 253 rows and 6 variables:
\describe{
\item{name}
\item{subj}
\item{x}
\item{y}
\item{delta}
\item{severe}
...
}
}
\usage{
real_dfs
}
\description{
Combined public dataframe

A dataset containing the information to do analysis on
correlations between x and y, and between x and delta
}
\keyword{datasets}
