% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate_data}
\usage{
simulate_data(
  n = 30,
  seed = 1,
  mu_x = 30,
  mu_y = 30,
  var_x = 100,
  var_ratio = 1,
  cor_xy = 0,
  empirical = TRUE,
  ...
)
}
\arguments{
\item{n}{sample size}

\item{seed}{seed for reproducibility}

\item{mu_x}{mean of the $x$ observations}

\item{mu_y}{mean of the $y$ observations}

\item{var_x}{variance of the $x$ observations}

\item{var_ratio}{variance ratio (\code{var_y} / \code{var_x})}

\item{cor_xy}{correlation betwene $x$ and $y$}

\item{empirical}{argument to \code{MASS::mvrnorm} -- should variances and correlations
be exactly as specified, or take these values on average?}

\item{...}{additional arguments}
}
\value{
dataframe containing simulated generated data
}
\description{
This function is used to simulate data illustrative of the
relationship between cor_xy, cor_xd, and the variance ratio. This function
is used to generate datasets A-E (dataset E is subjected to a ceiling
effect prior to analysis).
}
\examples{

simulate_data(n = 30, seed = 1,
             mu_x = 30, mu_y = 30,
            var_x = 100, var_ratio = 1, cor_xy = 0,
            empirical = TRUE)

}
