% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_proc.R
\name{gap_proc}
\alias{gap_proc}
\title{gap_proc}
\usage{
gap_proc(gap)
}
\arguments{
\item{gap}{an object produced by cluster::clusGap()}
}
\value{
a tidied dataframe containing the same results as the input object.
}
\description{
The clustering analyses make extensive use of the gap statistic, which is
implemented using cluster::clusGap(). The output of that function is
"untidy", and the purpose of this function is to clean output for subsequent
analysis and comparison.
}
